requirements.txt
Django>=4.2,<6.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.6
requests>=2.31.0
python-dotenv>=1.0.0
django-environ>=0.11.0
pytest-django>=4.5.2
requests-mock>=1.11.0
drf-spectacular>=0.26.0
django-filter>=23.2
ruff>=0.1.0
